==> notebooks/1b_trataABIMOTO13.ipynb <==
# _Notebook Jupyter_ 1b_trataABIMOTO13

# Importação e tratamento Tabela ABIMOTO v13

A tabela ABIMOTO v13 é fornecida em PDF, que foi manualmente transformado em uma planilha do Excel no formato XLSX.

### Importação das Bibliotecas

import pandas as pd
import numpy as np
import os, time

# Data e hora da execução do script
initot=time.time()
print(f'Código executado em {time.strftime("%d/%m/%Y às %H:%M", time.localtime(time.time()))}')

### Definindo dados gerais

Cria variáveis com nomes das colunas e seus tipos

tipos = { 'ITEM' : int, 
          'PARTES E PEÇAS' : str, 
          'MOTO PARTS' : str, 
          'NCM' : str, 
          'UNI' : str, 
          'VMLE' : float }
colunas = list(tipos.keys())

Inicializa a variável que conterá o tamanho total do dataset original

tamanhoDataset=0

Inicializa um dataframe vazio que conterá os dados finais

df = pd.DataFrame(columns = colunas)
df.head()

### Importa o arquivo Excel XLSX, trata e concatena no DataFrame final

# Executa a importação e tratamento do arquivo
arq = './bases/ABIMOTO13.xlsx'
print('Iniciando ' + arq + '.')
df = pd.read_excel(arq,
                   sep='@', 
                   decimal=r',', 
                   encoding = "ISO-8859-1",
                   header = 0,
                   names = colunas,
                   dtype = tipos, 
                   quotechar="'", 
                   error_bad_lines=False)
print('DataFrame carregado...\nAplicando filtros...')
# Elimina os registros sem valores ou nulos
df = df.dropna()
# Incrementa o tamanho do Dataset
tamanhoDataset += df[df.columns[0]].count()
print('Quantidade de registros válidos: ' + str(tamanhoDataset))
# Filtra o DataFrame somente com os registros de interesse
# Filtro 1: NCM de interesse: 87141000
indiceNCM = df['NCM'] == '87141000'
df = df[indiceNCM]
# Filtro 2: Descrição contendo 'kit' e 'transm'
df = df[df['PARTES E PEÇAS'].str.contains("kit", case=False)]
df = df[df['PARTES E PEÇAS'].str.contains("transm", case=False)]
print(arq + ' finalizado.\n')

### Ajustando a descrição e indicação de presença de corrente com retentor no kit

df.head()

Em virtude de haver observação se há ou não corrente com retentor na coluna UND faremos a adequação da descrição pra conter tal informação.

obs = ["COM RETENTOR" if "COM" in x else "SEM RETENTOR" for x in df.UNI.str.split()]
df["PARTES E PEÇAS"] = df["PARTES E PEÇAS"] + " " + obs

Em seguida define-se a coluna UNI como 'KIT'

df['UNI']='KIT'

E por fim cria-se a coluna RETENTOR do tipo _boolean_ para indicar ou não a presença do retentor no kit.

from funcoesTCC import retentorAux # função importada do módulo de funções criarModelo
df['RETENTOR'] = df['PARTES E PEÇAS'].apply(retentorAux)

### Resetando o índice do DataFrame importado

df.reset_index(inplace=True, drop=True)

### Excluindo colunas desnecessárias

excluir=['ITEM',
         'MOTO PARTS',
         'NCM',
         'UNI'] 

df=df.drop(excluir, axis=1)

### Verificando o DataFrame final

df.head(3)

df.tail(3)

df.shape

### Exportando o DataFrame

Exportando para um arquivo CSV

df.to_csv(r'./bases/dfABIMOTOv13.csv', index = False, header = True)

Exportando para um arquivo de planilha do Excel

df.to_excel(r'./bases/dfABIMOTOv13.xlsx', index = False, header = True)

Compara o tamanho total do Dataset inicial e final

print('Qtd de registros Dataset original: ' + str(tamanhoDataset))
print('Qtd de registros Dataset final:    ' + str(df[df.columns[0]].count()))

tempotot=time.time()-initot
if tempotot>60:
    print(f'Tempo total de execução: {tempotot/60:.2f} minutos.')
else:
    print(f'Tempo total de execução: {tempotot:.2f} segundos.')
