==> notebooks/4_classificarDESCRICAO.ipynb <==
# _Notebook Jupyter_ 4_classificarDESCRICAO

# Classificação da DESCRIÇÃO obtida a partir da lista de Aplicações

Diante da necessidade de se ter uma série de registros classificados para o aprendizado supervisionado será necessária a classificação prévia de toda a base de dados para a aplicação do aprendizado de máquina que fará predições futuras dessa classificação.

### Importa as bibliotecas necessárias

import pandas as pd, numpy as np
import time
# importa o CountVectorizer
from sklearn.feature_extraction.text import CountVectorizer

# Data e hora da execução do script
initot=time.time()
print(f'Código executado em {time.strftime("%d/%m/%Y às %H:%M", time.localtime(time.time()))}')

### Importando a lista de Aplicações

df_aplicacoes = pd.read_csv(r'./bases/Aplicacoes.csv')

df_aplicacoes.head()

df_aplicacoes.shape

Observa-se dos dados acima que há 861 modelos de motocicletas disponíveis, que são as nossas possíveis aplicações dos kits de transmissão a serem importados.

#### Criação das Bags of Words do DataSet de aplicações

bow_aplicacoes = CountVectorizer(token_pattern='(?u)\\b[a-zA-z0-9\\w-]+\\b')

display(bow_aplicacoes)

bow_aplicacoes.fit(df_aplicacoes['APLICACOES'])

# Verificação das palavras
vocabulario_aplicacoes = bow_aplicacoes.vocabulary_
print(str(vocabulario_aplicacoes)[:500]+' (...) '+str(vocabulario_aplicacoes)[-500:]) # amostra

# Verificação do número de termos no vocabulário de aplicações
len(bow_aplicacoes.vocabulary_)

# Transformação em vetores binários
X_bow_aplicacoes = bow_aplicacoes.fit_transform(df_aplicacoes['APLICACOES'])
print(f'{X_bow_aplicacoes[:2]}\n       (...)\n{X_bow_aplicacoes[-2:]}') # amostra

type(X_bow_aplicacoes)

matrix_aplicacoes = bow_aplicacoes.transform(df_aplicacoes['APLICACOES'])

print(f'{matrix_aplicacoes[:2]}\n       (...)\n{matrix_aplicacoes[-2:]}') # amostra

print(matrix_aplicacoes.shape)

print(matrix_aplicacoes.toarray())

df1=pd.DataFrame(matrix_aplicacoes.toarray())
df1.head()

# Totalização do número de termos de cada vocábulo
df_sum=df1.sum(axis=0)
df_sum.head()

# Sumário do dataset
df_sum.describe()

# Identificação dos valores máximos
print(df_sum.sort_values().head(),df_sum.sort_values().tail())

# Identificação dos valores com uma ocorrência
df_sum[df_sum==1].sample(5)

Diante da necessidade de se ter uma série de registros classificados para o aprendizado supervisionado e após a determinação dos 858 modelos de aplicações disponíveis, vamos classificar baseados nas seguintes observações realizadas na análise exploratória dos dados:
1. Os modelos de motocicletas têm nomes que os distinguem em grande parte dos casos;
2. Dos 913 termos, 655 aparacem apenas uma vez, o que implica que sua ocorrência já deve classificar o item;
3. O termo com maior número de ocorrências aparece 97 vezes (YAMAHA), sendo o menos distintivo de todos.

# Função para pegar a chave pelo valor, dado que valor é único.
def pegaChave(v):
    for chave, valor in vocabulario_aplicacoes.items():
        if v == valor:
            return chave
    return "Não existe chave para esse valor."

print(pegaChave(881).upper())

# Criação do dicionário chaves invertendo chave e valor do vocabulário
chaves = dict((v,k) for k,v in vocabulario_aplicacoes.items())

# Termo do vocabulário pelo índice
chaves[887].upper()

# Índice pelo termo do vocabulário
vocabulario_aplicacoes['yamaha']

# Número de ocorrências pelo termo do vocabulário
df_sum[vocabulario_aplicacoes['yamaha']]

# Função para determinar a aplicação caso a contagem do termo seja 1
def achaAplicacao1(modelo):
    # pega os termos do modelo e transforma em uma lista
    modelolst = modelo.lower().split()
    # para cada termo
    for mod in modelolst:
        try:
            # l é o índice termo no vocabulário de aplicações
            l = vocabulario_aplicacoes[mod]
            # se a soma de todos os termos do índice i do vocabulário for 1
            if df_sum[l]==1: 
                # pegar o índice dessa linha
                # que será onde o valor 1 aparece na coluna chamada linha do df1
                i = int(df1[df1[l]==1].index.values)
                # retornar a descrição da aplicação dessa linha
                return df_aplicacoes['APLICACOES'][i]
        except:
            continue
    return 'XXX'

print(achaAplicacao1('MOTO TEMOS OUTRO ZING'))

### Importa os dados já tratados

# Importa base de dados com os modelos já determinados para um dataframe
df = pd.read_excel('./bases/dataframe_modelos.xlsx')

df.sample(4)

df.iloc[:,-2:].head()

# Verifica o tamnanho do dataframe
df.shape

## Classificando segundo a lista de Aplicações

df['APLICACAO']=df['Modelo'].apply(achaAplicacao1)

df.iloc[:,-3:].head(10)

df[df['APLICACAO']=='XXX'].iloc[:,-3:].head(10)

print('Registros sem classificação:  ' + str(df[df['APLICACAO']=='XXX'].iloc[:,-3:].shape[0]))
print('Registros com classificação: ' + str(df[df['APLICACAO']!='XXX'].iloc[:,-3:].shape[0]))
print('Total de Registros:          ' + str(df.shape[0]))

Observa-se que após a classificação pelos termos únicos restaram cerca de 3.000 linhas, havendo mais de 15.000 registros já classificados.

#### Classifica o modelo como aplicação se os termos forem exatamente iguais

def achaAplicacao2(modelo):
    modelolst=modelo.lower().split() # cria lista com termos do modelo
    aplicacoes_temp=[] # inicializa a lista de saída
    for aplicacao in df_aplicacoes['APLICACOES']: # para cada aplicação
        aplicacaolst = aplicacao.lower().split() # cria lista
        if all(termos in aplicacaolst for termos in modelolst): # se a aplicacão contém o modelo
            aplicacoes_temp.append(aplicacao) # adiciona na lista de saída
    if len(aplicacoes_temp)==0:
        return 'XXX' # retorna a saída
    elif len(aplicacoes_temp)==1:
        return aplicacoes_temp[0] # retorna a saída
    else:
        return aplicacoes_temp # retorna a saída

# df temporário filtrado pelos não classificados
dftemp=df.iloc[:,-2:][df['APLICACAO']=='XXX']
dftemp.shape

dftemp.head()

# aplica a funão achaachaAplicacao2
dftemp=dftemp.assign(APLICACAO=dftemp['Modelo'].apply(achaAplicacao2))

dftemp.sample(5)

# quantidade de registros sem classificação
dftemp[dftemp['APLICACAO']=='XXX'].shape

# atualiza o dataframe df com as alterações feitas em dftemp
df.update(dftemp)

df.iloc[:,-3:].sample(5)

df[df['APLICACAO']=='XXX'].iloc[:,-3:].sample(5)

print('Registros sem classificação:  ' + str(df[df['APLICACAO']=='XXX'].iloc[:,-3:].shape[0]))
print('Registros com classificação: ' + str(df[df['APLICACAO']!='XXX'].iloc[:,-3:].shape[0]))
print('Total de Registros:          ' + str(df.shape[0]))

Observa-se que após a classificação pelos termos únicos restaram cerca de 600 linhas.<br>
Para fins de utilização no aprendizado de classificação, os quase 17.000 registros restantes são suficientes para comparação com os outros modelos e determinar a melhor classificação.

### Exportando o DataFrame

Exportando para um arquivo CSV

df.to_csv(r'./bases/dataframe_modelos_class0.csv', index = False, header = True)

Exportando para um arquivo de planilha do Excel

df.to_excel(r'./bases/dataframe_modelos_class0.xlsx', index = False, header = True)

tempotot=time.time()-initot
if tempotot>60:
    print(f'Tempo total de execução: {tempotot/60:.2f} minutos.')
else:
    print(f'Tempo total de execução: {tempotot:.2f} segundos.')
